backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database dependencies
motor>=3.3.2
pymongo>=4.6.0
beanie>=1.24.0

# Configuration and settings
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# Additional utilities
email-validator>=2.1.0
openai>=1.3.0
aiofiles>=23.2.1
requests>=2.31.0